Move `VisibilityFilter` to `shared`.
